Update common files
