+++ ui/app/src/models/Item.ts
@@ -13,9 +13,47 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-export interface Item {
+export interface LegacyItem {
   name: string;
   internalName: string;
   uri: string;
+
   [prop: string]: any;
+
+  children: LegacyItem[];
+}
+
+export interface LocalizationItem {
+  id: string;
+  label: string;
+  path: string;
+  previewUrl: string;
+  systemType: string;
+  mimeType: string;
+  state: number;
+  lockOwner: string;
+  disabled: boolean;
+  localeCode: string;
+  translationSourceId: string;
+}
+
+export interface Item {
+  id: string;
+  label: string;
+  contentTypeId: string;
+  path: string;
+  previewUrl: string;
+  systemType: string;
+  mimeType: string;
+  state: number;
+  lockOwner: string;
+  disabled: boolean;
+  localeCode: string;
+  translationSourceId: string;
+  creator: string;
+  createdDate: string;
+  modifier: string;
+  lastModifiedDate: string;
+  commitId: string;
+  sizeInBytes: number;
 }
