+++ ui/app/src/modules/Content/Publish/PublishDialog.tsx
@@ -15,17 +15,30 @@
  */
 
 import React, { PropsWithChildren, useCallback, useEffect, useMemo, useState } from 'react';
-import { defineMessages, useIntl } from 'react-intl';
+import { defineMessages, FormattedMessage, useIntl } from 'react-intl';
 import { fetchPublishingChannels } from '../../../services/content';
 import { goLive, submitToGoLive } from '../../../services/publishing';
 import { fetchDependencies } from '../../../services/dependencies';
-import PublishDialogUI from './PublishDialogUI';
-import { Item } from '../../../models/Item';
+import { LegacyItem } from '../../../models/Item';
 import moment from 'moment';
 import { useSelector } from 'react-redux';
 import GlobalState, { APIError } from '../../../models/GlobalState';
 import { useActiveSiteId, useOnMount, useSpreadState, useStateResource } from '../../../utils/hooks';
 import StandardAction from '../../../models/StandardAction';
+import { Resource } from '../../../models/Resource';
+import Grid from '@material-ui/core/Grid';
+import DependencySelection from '../Dependencies/DependencySelection';
+import PublishForm from './PublishForm';
+import { withStyles } from '@material-ui/core/styles';
+import Dialog from '@material-ui/core/Dialog';
+import DialogHeader from '../../../components/DialogHeader';
+import { CloseRounded } from '@material-ui/icons';
+import DialogBody from '../../../components/DialogBody';
+import { SuspenseWithEmptyState } from '../../../components/SystemStatus/Suspencified';
+import DialogFooter from '../../../components/DialogFooter';
+import Button from '@material-ui/core/Button';
+import CircularProgress from '@material-ui/core/CircularProgress';
+import { palette } from '../../../styles/theme';
 
 const goLiveMessages = defineMessages({
   title: {
@@ -61,7 +74,7 @@ export interface DependenciesResultObject {
   items2: []
 }
 
-export const checkState = (items: Item[]) => {
+export const checkState = (items: LegacyItem[]) => {
   return (items || []).reduce(
     (table: any, item) => {
       table[item.uri] = true;
@@ -85,7 +98,7 @@ export const updateCheckedList = (uri: string[], isChecked: boolean, checked: an
   return nextChecked;
 };
 
-export const selectAllDeps = (setChecked: Function, items: Item[]) => {
+export const selectAllDeps = (setChecked: Function, items: LegacyItem[]) => {
   setChecked(items.map(i => i.uri), true);
 };
 
@@ -95,10 +108,264 @@ export const paths = (checked: any) => (
     .map(([key]) => key)
 );
 
+const dialogStyles = () => ({
+  titleRoot: {
+    margin: 0,
+    padding: '13px 20px 11px',
+    background: palette.white
+  },
+  title: {
+    display: 'flex',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    paddingBottom: 10
+  },
+  subtitle: {
+    fontSize: '14px',
+    lineHeight: '18px',
+    paddingRight: '35px'
+  },
+  leftAlignedAction: {
+    marginRight: 'auto'
+  },
+  errorPaperRoot: {
+    maxHeight: '586px',
+    height: '100vh',
+    padding: 0
+  },
+  loadingStateRoot: {
+    height: '100%',
+  },
+  loadingStateGraphic: {
+    flexGrow: 1,
+    padding: '50px 0'
+  },
+  btnSpinner: {
+    marginLeft: 11,
+    marginRight: 11,
+    color: '#fff'
+  }
+});
+
+interface PublishDialogContentUIProps {
+  resource: Resource<any>;
+  checkedItems: LegacyItem[];
+  setCheckedItems: Function;
+  checkedSoftDep: any[];
+  setCheckedSoftDep: Function;
+  onClickSetChecked: Function;
+  deps: any,
+  showDepsButton: boolean;
+  selectAllDeps: Function;
+  selectAllSoft: Function;
+  dialog: any;
+  setDialog: any;
+  setSubmitDisabled: Function;
+  showEmailCheckbox?: boolean;
+  publishingChannelsStatus: string;
+  onPublishingChannelsFailRetry: Function;
+  apiState: any;
+}
+
+function PublishDialogContentUI(props: PublishDialogContentUIProps) {
+  const {
+    resource,
+    checkedItems,
+    setCheckedItems,
+    checkedSoftDep,
+    setCheckedSoftDep,
+    onClickSetChecked,
+    deps,
+    showDepsButton,
+    selectAllDeps,
+    selectAllSoft,
+    dialog,
+    setDialog,
+    setSubmitDisabled,
+    showEmailCheckbox,
+    publishingChannelsStatus,
+    onPublishingChannelsFailRetry,
+    apiState
+  } = props;
+
+  const { items, publishingChannels }: { items: LegacyItem[], publishingChannels: any } = resource.read();
+
+  return (
+    <>
+      <Grid container spacing={3}>
+        <Grid item xs={12} sm={7} md={7} lg={7} xl={7}>
+          <DependencySelection
+            items={items}
+            checked={checkedItems}
+            setChecked={setCheckedItems}
+            checkedSoftDep={checkedSoftDep}
+            setCheckedSoftDep={setCheckedSoftDep}
+            onClickSetChecked={onClickSetChecked}
+            deps={deps}
+            showDepsButton={showDepsButton}
+            onSelectAllClicked={selectAllDeps}
+            onSelectAllSoftClicked={selectAllSoft}
+            disabled={apiState.submitting}
+          />
+        </Grid>
+
+        <Grid item xs={12} sm={5} md={5} lg={5} xl={5}>
+          <PublishForm
+            inputs={dialog}
+            setInputs={setDialog}
+            setSubmitDisabled={setSubmitDisabled}
+            showEmailCheckbox={showEmailCheckbox}
+            publishingChannels={publishingChannels}
+            publishingChannelsStatus={publishingChannelsStatus}
+            onPublishingChannelsFailRetry={onPublishingChannelsFailRetry}
+            disabled={apiState.submitting}
+          />
+        </Grid>
+      </Grid>
+    </>
+  );
+}
+
+interface PublishDialogUIProps {
+  resource: Resource<any>;
+  publishingChannelsStatus: string;
+  onPublishingChannelsFailRetry: Function;
+  handleClose: any;
+  handleSubmit: any;
+  submitDisabled: boolean;
+  setSubmitDisabled: Function;
+  showDepsDisabled: boolean;
+  dialog: any;
+  setDialog: any;
+  open: boolean;
+  title: string;
+  subtitle?: string;
+  checkedItems: LegacyItem[];
+  setCheckedItems: Function;
+  checkedSoftDep: any[];
+  setCheckedSoftDep: Function;
+  onClickSetChecked: Function;
+  deps: any,
+  showDepsButton: boolean;
+  selectAllDeps: Function;
+  selectAllSoft: Function;
+  onClickShowAllDeps?: any;
+  showEmailCheckbox?: boolean;
+  apiState: any;
+  classes?: any;
+}
+
+const PublishDialogUI = withStyles(dialogStyles)((props: PublishDialogUIProps) => {
+  const {
+    resource,
+    publishingChannelsStatus,
+    onPublishingChannelsFailRetry,
+    handleClose,
+    handleSubmit,
+    submitDisabled,
+    setSubmitDisabled,
+    showDepsDisabled,
+    dialog,
+    setDialog,
+    open,
+    title,
+    subtitle,
+    checkedItems,
+    setCheckedItems,
+    checkedSoftDep,
+    setCheckedSoftDep,
+    onClickSetChecked,
+    deps,
+    showDepsButton,
+    selectAllDeps,
+    selectAllSoft,
+    onClickShowAllDeps,
+    showEmailCheckbox,
+    apiState,
+    classes
+  } = props;
+
+  return (
+    <Dialog
+      onClose={handleClose}
+      aria-labelledby="requestPublishDialogTitle"
+      open={open}
+      disableBackdropClick={true}
+      fullWidth={true}
+      maxWidth={'md'}
+    >
+      <DialogHeader
+        title={title}
+        subtitle={subtitle}
+        onClose={handleClose}
+        icon={CloseRounded}
+      />
+      <DialogBody>
+        <SuspenseWithEmptyState resource={resource}>
+          <PublishDialogContentUI
+            resource={resource}
+            checkedItems={checkedItems}
+            setCheckedItems={setCheckedItems}
+            checkedSoftDep={checkedSoftDep}
+            setCheckedSoftDep={setCheckedSoftDep}
+            onClickSetChecked={onClickSetChecked}
+            deps={deps}
+            showDepsButton={showDepsButton}
+            selectAllDeps={selectAllDeps}
+            selectAllSoft={selectAllSoft}
+            dialog={dialog}
+            setDialog={setDialog}
+            setSubmitDisabled={setSubmitDisabled}
+            showEmailCheckbox={showEmailCheckbox}
+            publishingChannelsStatus={publishingChannelsStatus}
+            onPublishingChannelsFailRetry={onPublishingChannelsFailRetry}
+            apiState={apiState}
+          />
+        </SuspenseWithEmptyState>
+      </DialogBody>
+      <DialogFooter>
+        <Button
+          color="primary"
+          onClick={ onClickShowAllDeps }
+          className={classes.leftAlignedAction}
+          disabled={showDepsDisabled || apiState.submitting}
+        >
+          <FormattedMessage
+            id="publishDialog.showAllDependencies"
+            defaultMessage={`Show All Dependencies`}
+          />
+        </Button>
+
+        <Button variant="contained" onClick={handleClose} disabled={apiState.submitting}>
+          <FormattedMessage
+            id="requestPublishDialog.cancel"
+            defaultMessage={`Cancel`}
+          />
+        </Button>
+        <Button variant="contained" autoFocus onClick={handleSubmit} color="primary" disabled={submitDisabled || apiState.submitting}>
+          {
+            apiState.submitting ?
+              (
+                <CircularProgress
+                  className={classes.btnSpinner}
+                  size={20}
+                />
+              ) : (
+                <FormattedMessage
+                  id="requestPublishDialog.submit"
+                  defaultMessage={`Submit`}
+                />
+              )
+          }
+        </Button>
+      </DialogFooter>
+    </Dialog>
+  )
+});
 
 interface PublishDialogBaseProps {
   open: boolean;
-  items?: Item[];
+  items?: LegacyItem[];
   scheduling?: string;
 }
